Fixing errors
Fixing the formatDate function to display the date correctly when the time difference is less than 24 hours.

Additionally, a critical error in the code responsible for saving messages in the database and sending them to participants in the conversation has been fixed.

Furthermore, improvements have been made to the README.md file to provide more precise information.